potentially wrong result when compare float with double

+++ YYKeyboardManager/YYKeyboardManager.m
@@ -245,11 +245,11 @@ static int _YYKeyboardViewFrameObserverKey;
 
 #pragma mark - private
 
-- (CGFloat)_systemVersion {
-    static CGFloat v;
+- (double)_systemVersion {
+    static double v;
     static dispatch_once_t onceToken;
     dispatch_once(&onceToken, ^{
-        v = [UIDevice currentDevice].systemVersion.floatValue;
+        v = [UIDevice currentDevice].systemVersion.doubleValue;
     });
     return v;
 }
